+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - 🐞 Bugfixes
 - 🔧 Others
 
+## TBD
+
+- ✨ Keep same `PipeStart` object throughout the pipe for improved performances
+
 ## 1.0.4 - 2024-11-22
 
 - ✨ Added `__slots__` to python implementation for improved performances