+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## TBD
 
 - ✨ [Python] Keep same `PipeStart` object throughout the pipe for improved performances
+- ✨ [Python]  the `PipeStart.history` attribute to keep track of all its values (only in debug mode)
 
 ## 1.0.4 - 2024-11-22